first version of runge_kutta4_implicit stepper using fixed point iterations
basically, this is just a slight adaption of the runge_kutta4_classic stepper. It inherits from explicit_stepper_base.
according to the one dimensional test equation

     x'   = x
     x(0) = 1

this stepper seems to be of fourth order if and only if the number of corrector steps is at least 3. If 2 iterations
steps are applied, the order is one less.

If no iterations are used, this stepper yields the same results as the explicit euler stepper.

tested for the state_types double and std::array<double,1>